Set Thread context classloder for entire middleware chain
