+++ src/main/java/com/microsoft/azure/functions/worker/broker/JavaMethodExecutor.java
@@ -1,14 +1,25 @@
 package com.microsoft.azure.functions.worker.broker;
 
-
-import java.util.*;
 import com.microsoft.azure.functions.worker.binding.*;
-
 
 /**
  * Used to executor of arbitrary Java method in any JAR using reflection.
  * Thread-Safety: Multiple thread.
  */
-public interface JavaMethodExecutor {
-    void execute(ExecutionContextDataSource executionContextDataSource) throws Exception;
+public class JavaMethodExecutor {
+
+    private static final JavaMethodExecutor INSTANCE = new JavaMethodExecutor();
+
+    public static JavaMethodExecutor getInstance(){
+        return INSTANCE;
+    }
+
+    private JavaMethodExecutor() {}
+
+    public void execute(ExecutionContextDataSource executionContextDataSource) throws Exception {
+        Object retValue = ParameterResolver.resolveArguments(executionContextDataSource)
+                .orElseThrow(() -> new NoSuchMethodException("Cannot locate the method signature with the given input"))
+                .invoke(executionContextDataSource::getFunctionInstance);
+        executionContextDataSource.updateReturnValue(retValue);
+    }
 }
